i8x41: add i8x41_scan().  m68k: revert int -> intptr_t change

+++ src/cpu/m68k/m68kcpu.c
@@ -1055,7 +1055,7 @@ unsigned int m68k_context_size()
 /* Used to calculate the context size minus the system-specific pointers, for savestates */
 unsigned int m68k_context_size_no_pointers()
 {
-	return (intptr_t)&m68ki_cpu.pointer_block_divider - (intptr_t)&m68ki_cpu;
+	return (int)&m68ki_cpu.pointer_block_divider - (int)&m68ki_cpu;
 }
 
 unsigned int m68k_get_context(void* dst)
